pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Chaisuu <Kfire13@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"


[tool.poetry.group.pythonProject.dependencies]
flake8 = "^7.1.0"
mypy = "^1.10.1"


[tool.poetry.group.pythonroject.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'


[tool.black]
# Максимальная длина строки
line-length = 119
exclude = '''
(
  /(
    | \.git
  )/

)
'''


[tool.isort]
# максимальная длина строки
line_length = 119
